Move command from `shopify app execute` to `shopify app bulk execute`
