+++ packages/app/src/cli/commands/app/bulk/execute.ts
@@ -1,13 +1,13 @@
-import {appFlags, bulkOperationFlags} from '../../flags.js'
-import AppLinkedCommand, {AppLinkedCommandOutput} from '../../utilities/app-linked-command.js'
-import {linkedAppContext} from '../../services/app-context.js'
-import {storeContext} from '../../services/store-context.js'
-import {executeBulkOperation} from '../../services/bulk-operations/execute-bulk-operation.js'
+import {appFlags, bulkOperationFlags} from '../../../flags.js'
+import AppLinkedCommand, {AppLinkedCommandOutput} from '../../../utilities/app-linked-command.js'
+import {linkedAppContext} from '../../../services/app-context.js'
+import {storeContext} from '../../../services/store-context.js'
+import {executeBulkOperation} from '../../../services/bulk-operations/execute-bulk-operation.js'
 import {globalFlags} from '@shopify/cli-kit/node/cli'
 import {readStdinString} from '@shopify/cli-kit/node/system'
 import {AbortError} from '@shopify/cli-kit/node/error'
 
-export default class Execute extends AppLinkedCommand {
+export default class BulkExecute extends AppLinkedCommand {
   static summary = 'Execute bulk operations.'
 
   static description = 'Execute bulk operations against the Shopify Admin API.'
@@ -21,13 +21,13 @@ export default class Execute extends AppLinkedCommand {
   }
 
   async run(): Promise<AppLinkedCommandOutput> {
-    const {flags} = await this.parse(Execute)
+    const {flags} = await this.parse(BulkExecute)
 
     const query = flags.query ?? (await readStdinString())
     if (!query) {
       throw new AbortError(
         'No query provided. Use the --query flag or pipe input via stdin.',
-        'Example: echo "query { ... }" | shopify app execute',
+        'Example: echo "query { ... }" | shopify app bulk execute',
       )
     }
